sr(api): rename user-tutorial-serializer to user-saved-tutorial-serializer
