+++ api/tests/acceptance/application/tutorials/user-tutorials-controller_test.js
@@ -359,6 +359,12 @@ describe('Acceptance | Controller | user-tutorial-controller', function () {
                   type: 'user-saved-tutorial',
                 },
               },
+              'user-tutorial': {
+                data: {
+                  id: `${userSavedTutorialId}`,
+                  type: 'user-tutorial',
+                },
+              },
             },
             type: 'tutorials',
           },
@@ -379,6 +385,9 @@ describe('Acceptance | Controller | user-tutorial-controller', function () {
               'user-saved-tutorial': {
                 data: null,
               },
+              'user-tutorial': {
+                data: null,
+              },
             },
             type: 'tutorials',
           },
@@ -397,6 +406,9 @@ describe('Acceptance | Controller | user-tutorial-controller', function () {
                 data: null,
               },
               'user-saved-tutorial': {
+                data: null,
+              },
+              'user-tutorial': {
                 data: null,
               },
             },
@@ -466,6 +478,9 @@ describe('Acceptance | Controller | user-tutorial-controller', function () {
               'user-saved-tutorial': {
                 data: null,
               },
+              'user-tutorial': {
+                data: null,
+              },
             },
             type: 'tutorials',
           },
@@ -484,6 +499,9 @@ describe('Acceptance | Controller | user-tutorial-controller', function () {
                 data: null,
               },
               'user-saved-tutorial': {
+                data: null,
+              },
+              'user-tutorial': {
                 data: null,
               },
             },
@@ -677,6 +695,7 @@ describe('Acceptance | Controller | user-tutorial-controller', function () {
             },
             relationships: {
               'user-saved-tutorial': { data: { id: '101', type: 'user-saved-tutorial' } },
+              'user-tutorial': { data: { id: '101', type: 'user-tutorial' } },
               'tutorial-evaluation': {
                 data: null,
               },
@@ -751,6 +770,12 @@ describe('Acceptance | Controller | user-tutorial-controller', function () {
                   type: 'user-saved-tutorial',
                 },
               },
+              'user-tutorial': {
+                data: {
+                  id: '103',
+                  type: 'user-tutorial',
+                },
+              },
             },
             type: 'tutorials',
           },
@@ -772,6 +797,12 @@ describe('Acceptance | Controller | user-tutorial-controller', function () {
                 data: {
                   id: '102',
                   type: 'user-saved-tutorial',
+                },
+              },
+              'user-tutorial': {
+                data: {
+                  id: '102',
+                  type: 'user-tutorial',
                 },
               },
             },
